Finally fixed issue with extra columns
I didn't necessarily fix it, I couldn't find anything online on why that was occurring and I did try multiple different things with no success. So what I ended up doing was creating a function to clean the dataframe and remove extra columns so that the data is cleaned. It's not pretty, doesn't solve the cause, but it ends up doing what I want anyways in O(n) time