Refined algorithm to find company symbol
Before, the company symbol was pretty inaccurate, it would search for names and other things it mentioned. So I made it so that it would have to keep searching until the first word company it talked about was in the article at least twice and the symbols it wants to check for must be in US currency. After doing that, I started seeing a lot more accurate company symbols being saved