Using financial lexicon from University of Notre Dame
Using https://sraf.nd.edu code and lexicon to analyze words. The issue I was running into was that the NLTK would analyze words but in a general sense and not financially. So by using their lexicon, I can make the article analysis more accurate financially. Hopefully this may fix the distrubution of scores in the data and break it up more instead of having the large clusters of data around the positive