Created NGram counts per frequency
I looked at the ngrams of each article and the difference in price between release date and date of the next article. For every positive difference, I added to the increase column for every ngram and every decrease column. So now I can look at an ngram and see how many times it shows up in positive articles and negative articles too